+++ Realm/RLMObjectBase.h
@@ -133,4 +133,31 @@ NS_ASSUME_NONNULL_BEGIN
 @interface RealmSwiftEmbeddedObject : RLMObjectBase
 @end
 
+/**
+  `RLMAsymmetricObject` is a base class used to define asymmetric Realm model objects.
+
+ Asymmetric objects can only be created using the `createObject:`
+ function, and cannot be added, removed or queried.
+ When created asymmetric objects will synced unidirectionally to the  MongoDB
+ database and cannot be accessed locally.
+
+ Incoming links from and asymmetric table are not allowed, meaning embedding
+ an asymmetric object within an `RLMObject` will throw an error.
+
+ The property types supported on `RLMAsymmetricObject` are the same as for `RLMObject`,
+ except for that asymmetric objects can only link to embedded objects, so `RLMObject`
+ and `RLMArray<RLMObject>` properties are not supported (`RLMEmbeddedObject` and
+ `RLMArray<RLEmbeddedObject>` *are*).
+
+ ```swift
+ class Person: AsymmetricObject {
+     @Persisted(primaryKey: true) var _id: ObjectId = ObjectId.generate()
+     @Persisted var name: String
+     @Persisted var age: Int
+ }
+ ```
+ */
+@interface RealmSwiftAsymmetricObject : RLMObjectBase
+@end
+
 NS_ASSUME_NONNULL_END