+++ Realm/RLMObjectBase.mm
@@ -356,6 +356,10 @@ id RLMCreateManagedAccessor(Class cls, RLMClassInfo *info) {
     return false;
 }
 
++ (bool)isAsymmetric {
+    return false;
+}
+
 - (id)mutableArrayValueForKey:(NSString *)key {
     id obj = [self valueForKey:key];
     if ([obj isKindOfClass:[RLMArray class]]) {
@@ -785,6 +789,12 @@ uint64_t RLMObjectBaseGetCombineId(__unsafe_unretained RLMObjectBase *const obj)
 @end
 
 @implementation RealmSwiftEmbeddedObject
++ (BOOL)accessInstanceVariablesDirectly {
+    return NO;
+}
+@end
+
+@implementation RealmSwiftAsymmetricObject
 + (BOOL)accessInstanceVariablesDirectly {
     return NO;
 }