+++ Realm/RLMObjectStore.h
@@ -61,6 +61,10 @@ RLMObjectBase *RLMCreateObjectInRealmWithValue(RLMRealm *realm, NSString *classN
                                                id _Nullable value, RLMUpdatePolicy updatePolicy)
 NS_RETURNS_RETAINED;
 
+// creates an asymmetric object and doesn't return
+void RLMCreateAsymmetricObjectInRealm(__unsafe_unretained RLMObjectBase *const object,
+                                      __unsafe_unretained RLMRealm *const realm);
+
 //
 // Accessor Creation
 //